+++ python/aistore/sdk/obj/content_iterator/sequential.py
@@ -4,11 +4,11 @@
 
 from typing import Generator
 
-from aistore.sdk.obj.object_client import ObjectClient
+from aistore.sdk.obj.content_iterator.base import BaseContentIterProvider
 
 
 # pylint: disable=too-few-public-methods
-class ContentIterProvider:
+class ContentIterProvider(BaseContentIterProvider):
     """
     Provide an iterator to open an HTTP response stream and read chunks of object content.
 
@@ -16,20 +16,6 @@ class ContentIterProvider:
         client (ObjectClient): Client for accessing contents of an individual object.
         chunk_size (int): Size of each chunk of data yielded from the response stream.
     """
-
-    def __init__(self, client: ObjectClient, chunk_size: int):
-        self._client = client
-        self._chunk_size = chunk_size
-
-    @property
-    def client(self) -> ObjectClient:
-        """
-        Get the client associated with this content iterator.
-
-        Returns:
-            ObjectClient: The client used to access object content.
-        """
-        return self._client
 
     def create_iter(self, offset: int = 0) -> Generator[bytes, None, None]:
         """